refactored to make things a bit simpler in preparation for parser overhaul
